requirements.txt
cadquery
requests
pillow